requirements.txt
breeze-connect>=1.0.65
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
yfinance>=0.2.30
python-dotenv>=1.0.0
